+++ lib/Venus.pm
@@ -7,7 +7,7 @@ use warnings;
 
 # VERSION
 
-our $VERSION = '1.85';
+our $VERSION = '1.90';
 
 # AUTHORITY