+++ CHANGES.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version: 1.90
+
+- [feature] Allow option_build to set value on read/write
+- [feature] Allow option_build on object construction
+- [feature] Test permutations of option_build, option_default, and option_coerce
+- [feature] Implement lazy_build hook in Venus::Role::Optional
+- [feature] Implement self_coerce hook in Venus::Role::Optional
+- [feature] Implement self_assert hook in Venus::Role::Optional
+- [feature] Use Venus::Assert#expression when asserting with option_assert
+
 ## Version: 1.85
 
 - [feature] Implement Venus::Fault